Adding a command line argument to be able to pass ingest_url
Recently a change went in that validates the API token that is passed
as an argument against prod. But given that we use the shell install
script for testing against lab, we needed an option to pass non-prod
urls.
This change is to purely help internal use cases and testing.